requirements.txt
python-dotenv
spotipy
pandas
